+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+...
+
+## [0.2.0] - 2019-05-11
+
+This driver is now functionally complete.
+
 ### Added
 - Enable/disable MEMS self-test function.
 - Support interrupts.
@@ -24,5 +30,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 This is the initial release to crates.io. All changes will be documented in
 this CHANGELOG.
 
-[Unreleased]: https://github.com/eldruin/kxcj9-rs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/eldruin/kxcj9-rs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/eldruin/kxcj9-rs/compare/v0.1.0...v0.2.0